[story/TP-95093]: Split tasks.rb into multiple class files and directories to make it easier to follow what the tasks are actually doing. Add a thor file to load all the files and update docs to specify installing the directory rather than the old tasks.rb file.
