+++ README.md
@@ -54,7 +54,7 @@ Installing the DataMapper thor tasks
 internet, you can also clone the github repo containing the tasks and
 then install them like you would install any thor task.~~
 
-    thor install dm-dev/tasks.rb
+    thor install dm-dev/tasks
 
 Either way, after showing you the content of tasks.rb, thor will ask you for a
 name for those new tasks.
