+++ tasks/app.rb
@@ -1,0 +1,7 @@
+#!/usr/bin/env ruby -Ku
+require 'English'
+require 'io/console'
+
+require_relative 'datamapper/project'
+
+DATAMAPPER_PROJECT = DataMapper::Project
