+++ tasks/main.thor
@@ -1,0 +1,3 @@
+$LOAD_PATH << File.expand_path(__dir__.to_s)
+
+require_relative 'app'
