+++ tasks/project/command/bundle/list.rb
@@ -1,0 +1,14 @@
+require_relative '../rvm/exec'
+require_relative '../bundle'
+
+class ::Project
+  class Command
+    class Bundle < Rvm::Exec
+      class List < Bundle
+        def bundle_command
+          'list'
+        end
+      end
+    end
+  end
+end
