+++ tasks/project/command/list.rb
@@ -1,0 +1,11 @@
+require_relative '../command'
+
+class ::Project
+  class Command
+    class List < ::Project::Command
+      def run
+        log
+      end
+    end
+  end
+end
