+++ tasks/project/command/release.rb
@@ -1,0 +1,26 @@
+require 'fileutils'
+require_relative '../command'
+
+class ::Project
+  class Command
+    class Release < Command
+      def run
+        # TODO: move to its own command
+        clean_repository(project_name)
+
+        FileUtils.cd(working_dir) do
+          log(command)
+          system(command) unless pretend?
+        end
+      end
+
+      def command
+        'rake release'
+      end
+
+      def action
+        'Releasing'
+      end
+    end
+  end
+end
